chore(eslint): zero gradual config on src
